service: http: util: testing: Move testing infra
